# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Comparador de Preços API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ofertas:
    get:
      operationId: listOfertas
      tags: [ofertas]
      summary: List offers
      parameters:
        - name: produto
          in: query
          schema:
            type: string
          description: Filter by product name
        - name: categoria
          in: query
          schema:
            type: string
          description: Filter by category
        - name: lat
          in: query
          schema:
            type: number
          description: User latitude
        - name: lng
          in: query
          schema:
            type: number
          description: User longitude
        - name: raio
          in: query
          schema:
            type: number
          description: Radius in kilometers
        - name: ordenar
          in: query
          schema:
            type: string
            enum: [preco, distancia, validacoes, recente, score]
          description: Sort order
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Oferta"
    post:
      operationId: createOferta
      tags: [ofertas]
      summary: Create a new offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfertaBody"
      responses:
        "201":
          description: Created offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oferta"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ofertas/{id}/like:
    post:
      operationId: likeOferta
      tags: [ofertas]
      summary: Like an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionBody"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oferta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ofertas/{id}/validar:
    post:
      operationId: validarOferta
      tags: [ofertas]
      summary: Validate an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionBody"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oferta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ofertas/{id}/denunciar:
    post:
      operationId: denunciarOferta
      tags: [ofertas]
      summary: Report a suspicious offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionBody"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oferta"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Oferta:
      type: object
      properties:
        id:
          type: integer
        produto:
          type: string
        categoria:
          type: string
        marca:
          type: string
          nullable: true
        preco:
          type: number
        mercado:
          type: string
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        fotoUrl:
          type: string
          nullable: true
        validade:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        dataCriacao:
          type: string
        curtidas:
          type: integer
        validacoes:
          type: integer
        denuncias:
          type: integer
        status:
          type: string
          enum: [nova, validada, suspeita, expirada]
        usuario:
          type: string
        usuarioId:
          type: integer
        ultimaValidacaoEm:
          type: string
          nullable: true
        score:
          type: integer
        nivelUsuario:
          type: string
          enum: [Iniciante, Explorador, "Caçador", Especialista, Mestre, Lenda]
        distancia:
          type: number
          nullable: true
        destacada:
          type: boolean
        patrocinada:
          type: boolean
        confirmacoes:
          type: integer
        ultimaConfirmacaoEm:
          type: string
          nullable: true
        wasConfirmation:
          type: boolean
          description: "Present only in POST /ofertas response: true when the submission confirmed an existing offer instead of creating a new one."
      required:
        - id
        - produto
        - categoria
        - preco
        - mercado
        - dataCriacao
        - curtidas
        - validacoes
        - denuncias
        - status
        - usuario
        - usuarioId
        - score
        - nivelUsuario
        - destacada
        - patrocinada
        - confirmacoes
...
    ActionBody:
      type: object
      required:
        - usuarioId
      properties:
        usuarioId:
          type: integer
          description: ID of the authenticated user performing the action
    CreateOfertaBody:
      type: object
      properties:
        produto:
          type: string
        categoria:
          type: string
        marca:
          type: string
          nullable: true
        preco:
          type: number
        mercado:
          type: string
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        fotoUrl:
          type: string
          nullable: true
        validade:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        usuarioId:
          type: integer
      required:
        - produto
        - categoria
        - preco
        - mercado
        - usuarioId
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error